refactor(b): organize `structure`
